migrate Indian Rupee Sign check to the `universal` profile
(issue #2969)
