+++ tests/profiles/googlefonts_test.py
@@ -3485,24 +3485,3 @@ def test_check_varfont_unsupported_axes():
   assert_results_contain(check(ttFont),
                          FAIL, 'unsupported-slnt')
 
-
-def test_check_rupee():
-  """ Ensure indic fonts have the Indian Rupee Sign glyph. """
-  from fontbakery.profiles.googlefonts import (com_google_fonts_check_rupee as check,
-                                               is_indic_font)
-
-  # This one is good:
-  ttFont = TTFont("data/test/indic-font-with-rupee-sign/NotoSerifDevanagari-Regular.ttf")
-  assert_PASS(check(ttFont),
-              "with a sample font that has the Indian Rupee Sign.")
-
-  # But this one lacks the glyph:
-  ttFont = TTFont("data/test/indic-font-without-rupee-sign/NotoSansOlChiki-Regular.ttf")
-  assert_results_contain(check(ttFont),
-                         FAIL, "missing-rupee",
-                         "with a sample font missing it.")
-
-  print("Ensure the check will SKIP when dealing with a non-indic font...")
-  non_indic = TTFont("data/test/mada/Mada-Regular.ttf")
-  assert is_indic_font(non_indic) == False
-
